Refactor CutlassFusedMoE
